requirements.txt
dash==2.11.1
dash-bootstrap-components==1.4.2
dash_daq==0.5.0
dash_trich_components==1.0.0
pandas==2.0.3
numpy==1.25.1
plotly==5.15.0
python_dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
dash-tools
openpyxl==3.0.10
dash-breakpoints==0.1.0
pyarrow==11.0.0
dash-extensions==1.0.4
celery==5.3.6
diskcache==5.6.1
multiprocess>=0.70.12
psutil>=5.8.0
flower==1.0.0
cached-property==1.5.2
packaging==21.3
pyparsing==3.0.8
pip==23.0.1
py-mini-racer==0.6.0
asyncio==3.4.3
dash-loading-spinners==1.0.0
redis==4.6.0
polars==0.18.4
dash[celery]==2.11.1
Flask==2.2.5
Flask-Caching==2.0.2
Werkzeug<2.3.7